robot/pyproject.toml
[tool.poetry]
name = "robot"
version = "0.1.0"
description = ""
authors = ["Noah May <noahmouse2011@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mistune = "^0.8.4"
python-slugify = "^4.0.0"
bs4 = "^0.0.1"
requests = "^2.22.0"
pyinstall = "^0.1.4"
pyinstaller = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
